requirements.txt
# requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
pydantic
python-multipart
